# Add LRU generics using code generation.