--- a/tx_applier.go
+++ b/tx_applier.go
@@ -21,8 +21,10 @@ package wavelet
 
 import (
 	"encoding/hex"
+
 	wasm "github.com/perlin-network/life/wasm-validation"
 	"github.com/perlin-network/wavelet/avl"
+
 	"github.com/perlin-network/wavelet/log"
 	"github.com/perlin-network/wavelet/sys"
 	"github.com/pkg/errors"