The library [https://github.com/cheekybits/genny](https://github.com/cheekybits/genny) is used for the code generation.

To generate the code, run `go generate`. It will generate 4 output files containing 4 implementations of LRU.

The output files will be overwritten, so it's safe to call `go generate` many times.

Following are the results from simple benchmarks comparing the `interface{}` LRU to the generics LRU.

```
name       old time/op    new time/op    delta
LRULoad-8    74.9ns ± 2%    63.3ns ± 9%  -15.47%  (p=0.000 n=88+100)

name       old alloc/op   new alloc/op   delta
LRULoad-8     0.00B          0.00B          ~     (all equal)

name       old allocs/op  new allocs/op  delta
LRULoad-8      0.00           0.00          ~     (all equal)
```

<br />

```
name      old time/op    new time/op    delta
LRUPut-8     127ns ± 7%      62ns ± 6%   -51.28%  (p=0.000 n=89+97)

name      old alloc/op   new alloc/op   delta
LRUPut-8     96.0B ± 0%      0.0B       -100.00%  (p=0.000 n=100+100)

name      old allocs/op  new allocs/op  delta
LRUPut-8      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=100+100)
```

<br />

```
name            old time/op    new time/op    delta
LRULoadOrPut-8     126ns ± 3%      65ns ± 3%   -48.49%  (p=0.000 n=98+90)

name            old alloc/op   new alloc/op   delta
LRULoadOrPut-8     96.0B ± 0%      0.0B       -100.00%  (p=0.000 n=100+100)

name            old allocs/op  new allocs/op  delta
LRULoadOrPut-8      2.00 ± 0%      0.00       -100.00%  (p=0.000 n=100+100)
```